MDL-26504 blog: made it delete external blog posts when the external blog is deleted

+++ blog/external_blogs.php
@@ -44,7 +44,15 @@ $message = null;
 if ($delete && confirm_sesskey()) {
     $externalbloguserid = $DB->get_field('blog_external', 'userid', array('id' => $delete));
     if ($externalbloguserid == $USER->id) {
+        // Delete the external blog
         $DB->delete_records('blog_external', array('id' => $delete));
+
+        // Delete the external blog's posts
+        $deletewhere = 'module = :module
+                            AND userid = :userid
+                            AND ' . $DB->sql_compare_text('content') . ' = ' . $DB->sql_compare_text($delete);
+        $DB->delete_records_select('post', $deletewhere, array('module' => 'blog_external', 'userid' => $USER->id));
+
         $message = get_string('externalblogdeleted', 'blog');
     }
 }
